Ideal Case
Explanation:

This VHDL code defines an entity called ideal_case with 8 input signals (input1 through input8) and 8 output signals (output1 through output8). The architecture of this entity implements logic functions using various logic gates (AND, OR, NOT) to compute the values of the output signals based on the input signals.